Switch to 3.7.0-rc2 with working console.

+++ arch/m68k/sun3x/time.c
@@ -71,7 +71,7 @@ int sun3x_hwclk(int set, struct rtc_time *t)
 	return 0;
 }
 /* Not much we can do here */
-u32 sun3x_gettimeoffset(void)
+unsigned long sun3x_gettimeoffset (void)
 {
     return 0L;
 }
